Added a new Makefile for building guess in-place.
The new Makefile compiles the files in the file hierarchy they have in the
repository, so there's no need to move the files to a separate build directory.
The results are placed in a new directory named "output". The Makefile can
build both the commandline version and the Simba PBS version.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/guess2008@14 70805675-57b6-444b-bb00-18ecffcb4be8