Removed output directory as separate target in Makefile.
This was done because it seemed to cause problems with unnecessary rebuilding
(since adding new results to the output directory modified it, which caused
other targets to need rebuilding etc. etc.) Also it doesn't seem to be
necessary since the creation of the .o files creates the output directory
anyway.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/guess2008@28 70805675-57b6-444b-bb00-18ecffcb4be8