Added the searchradius parameter to the CRU I/O module.
By specifying the searchradius parameter in the instruction file, CRU 
coordinates will be searched for in a circle if the ones given in the gridlist
don't match anything in the CRU data set.

If searchradius isn't specified, or set to zero, only the exact coordinates
given in the gridlist will be used.

This functionality was already implemented in the EMDI benchmarks, but now
it's a bit more general and by having it in the standard code the EMDI 
benchmarks need less special code.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/guess2008@29 70805675-57b6-444b-bb00-18ecffcb4be8