Renamed main.cpp and main.h in windows_version to dllmain.
The reason for this is that old Visual Studio 6 places all .obj files in a workspace in a single directory (even if there are multiple projects). So if command_line_version/main.cpp is in one project and windows_version/main.cpp is in another (in the same workspace) they will collide.

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/guess2008@60 70805675-57b6-444b-bb00-18ecffcb4be8