Changed how the post processing of benchmarks works a little.
Instead of having a common_postprocess.sh which runs for all benchmarks there
is now a folder for scripts which can be called by all benchmarks. The stuff
which used to be in common_postprocess.sh was moved to this folder and the
benchmarks that want to include that in their post processing can start their
postprocess.sh by running common1961to1990.sh.

Benchmarks which shouldn't have any post processing can simply omit the
postprocess.sh file.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/guess2008@78 70805675-57b6-444b-bb00-18ecffcb4be8