Bug fix: the fire probability didn't work properly in the euroflux benchmarks.
The fire functionality was switched off in the euroflux benchmarks after
the plantation date, however the Patch::fireprob variable wasn't updated after
that so it kept the same value as for the previous year. Now it is set to
zero if there is no fire.

Also made sure that fireprob is initialized to 0 when a new Patch object is
created.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/tech@158 70805675-57b6-444b-bb00-18ecffcb4be8