Fixed a bug in the post processing of most benchmarks (relating to gmapall).
The gmapall commands in common1961to1990.sh were meant to create maps for
lai and cmass for all species/PFTs. However the cmass pictures overwrote the
lai pictures. Now a new version of gmapall is used instead which has an option
for giving the picture filenames a prefix so that they don't collide.

Copied to gmapall script from the utilities package to the postprocessing
directory for now to avoid problems with compatibility issues.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/benchmark_improvements@179 70805675-57b6-444b-bb00-18ecffcb4be8