Some preparations for allowing ins-files to include other ins-files.
The submit scripts now doesn't copy the ins file to each run directory, instead
the absolute path to the ins file is used when starting guess.

Also the benchmarks script sends in the ins file as a parameter to the submit
script.

This way the ins files can remain in their original position and have relative
paths to other ins files in that same directory structure.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/tech@241 70805675-57b6-444b-bb00-18ecffcb4be8