Refactored photosynthesis and assimilation_wstress a bit and fixed two bugs.
photosynthesis and assimilation_wstress now return a PhotosynthesisResult object instead of several doubles. This also means there is only one if these functions, no special bvoc-version.

The bug fixes were in assimilation_wstress, in the first if statement the whole PhotosynthesisResult object is now cleared, before only assim was set to zero and the rest undefined. The other bug fix was that the whole PhotosynthesisResult is now included in the lookup table assimilation_wstress uses.

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/BVOC@300 70805675-57b6-444b-bb00-18ecffcb4be8