Bug fix in forest_floor_conditions, makes the model run faster.
This shouldn't alter the results in any way, but can have a fairly large inpact on speed.

The forest_floor_conditions function calculates photosynthesis for each Standpft, but forgot to set the flag to mark that it had been calculated, so it was calculated needlessly over and over.

The reason for adding the calculation of Standpft::gpterm becomes clear if looking at the similar photosynthesis calculation in demand().


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/bugfix@317 70805675-57b6-444b-bb00-18ecffcb4be8