Changed the order of multiplications at one place in forest_floor_conditions.
This brings it back to the order it was done before r300, which seems to have introduced differences in results for some grid cells.

Not sure which would give better precision, but for now I changed it so it would be easier to compare results for the BVOC branch with the results from trunk.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/BVOC@323 70805675-57b6-444b-bb00-18ecffcb4be8