The euroflux output file eurofluxstats.out was split into three files (nee,aet and gpp).
This makes the observed values behave like the regular monthly output files mnee, maet and mgpp.

The observed values are now printed with the same units as in the regular monthly output files, which means that the plots produced by the post processing is slightly changed. The units have changed for the nee and gpp plots, for nee the values have also been negated which caused the plot to be flipped around.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/tech@352 70805675-57b6-444b-bb00-18ecffcb4be8