First version of new output system.
The new OutputChannel system separates the formatting of output from the outannual function. This way formatting can change, or output can be sent to something other than files without making big changes in the I/O module. This also simplifies outannual.

The output for emdi and euroflux changes a little because of this, but the changes should only be formatting changes (white space etc).


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/tech@359 70805675-57b6-444b-bb00-18ecffcb4be8