Some changes to the benchmarks reporting.
Images are now not resized, instead they are by default links. Small images can be embedded in the html page with an extra parameter to describe_image.

Removed the image resizing to get rid of the dependency on ImageMagick, which isn't installed by default on compute nodes in (probably) most clusters.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/benchmark_improvements@378 70805675-57b6-444b-bb00-18ecffcb4be8