Bugfix from Veiko to prevent infinite loop in prdaily.
The problem was that when the monthly precipitation values were very small,
but not considered negligible by the neglibigle() function, the 
while(negligible(mprec_sum)) loop could go on forever.

Now instead we check if mval_prec is less than 0.1.

Quote from Veiko:

You are then not distributing rain if it is below 1/10 mm which is well 
below the dew you might get every morning. So the error is negligible.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/bugfix@487 70805675-57b6-444b-bb00-18ecffcb4be8