Support for NetCDF in the CMake build system.
CMake will now try to find NetCDF libraries and include files and use them if
available.

If NetCDF isn't installed the user can specify that for instance a local 
installation in the user's home directory should be used. This is done either
by setting the environment variable CMAKE_PREFIX_PATH before running CMake, or
by switching to advanced mode in the CMake interface and specifying each path.

The C++ API currently doesn't seem to be properly supported for Windows, and 
judging by the Unidata web site it seems they might switch to a newer C++ API
eventually, so if the C++ library isn't found NetCDF is still considered to be
found, but only the C library is then available.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/tech@516 70805675-57b6-444b-bb00-18ecffcb4be8