Added "submit" script for running GUESS in parallel on multi-core systems.
The script works like the submit scripts for the clusters, but parallelizes
and runs the simulation locally.

The variables controlling the run (NPROCESSES, GRIDLIST etc.) work in the
same way (WALLTIME not supported though).

Results are appended automatically, but only if all processes finish with
exit code 0 (success). If the user hits Ctrl-C all processes are aborted.

To use, set SYSTEM to multicore in cmake.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/tech@1080 70805675-57b6-444b-bb00-18ecffcb4be8