Moved the framework() function into its own file (framework.h/framework.cpp).
This lets us have method implementations for the classes in guess.h in 
guess.cpp without hiding the framework function below a lot of irrelevant code.

Since the framework function is such an important function for anyone getting
to know the model, it's nice if it's accessible and easy to read. When the
save state-feature gets in place there will be a need for many more methods
which should be implemented in the guess.cpp file.

This also prepares a little for breaking out the part of framework which
simulates a single day, and adding code for running all gridcells concurrently
or calling this code from a coupled model.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/tech@1089 70805675-57b6-444b-bb00-18ecffcb4be8