GlobalCO2File now fails if CO2 values after the last year are requested.
This was suggested by Stefan Olin, to make sure we don't accidentaly run
with bad CO2 values and don't notice it.

Needed to change the getclimate function slightly also to terminate earlier
once we've passed the historical period, to make sure we never query the co2
object for years after the historical period.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/tech@1105 70805675-57b6-444b-bb00-18ecffcb4be8