Took out implementation of dprintf, fail etc from main.cpp for more flexibility.
The functions remain and work like before, but behind the scenes a global 
"shell" object is now used to actually do the work. This shell object can
have different implementations depending on where you want to send messages
to the user for instance (there's one implementation for the command line
client, and one for the Windows shell, but others could be added).

This makes it possible to compile the model without main.cpp, for instance
when coupling LPJ-GUESS to another model as a library, or for unit testing.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/tech@1110 70805675-57b6-444b-bb00-18ecffcb4be8