Added a lightweight framework for unit tests, and some trivial tests for Date.
The CATCH unit test framework is used for now since it is very lightweight and
simple to integrate into the build process (everything is in a single header).

Unit tests are by default not compiled for now. If UNIT_TESTS is switched on
in CMake, they are compiled when LPJ-GUESS is built, and runs automatically as 
a post build step.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/tech@1112 70805675-57b6-444b-bb00-18ecffcb4be8