Support for insolation as radiation during whole time period (not just daytime).
Now there's no need to modify daylengthinsoleet when using a climate data set
with insolation expressed as radiation during each whole time period. Instead
a different insoltype is chosen by the I/O module.

Also tried to make the documentation about this clear as it seems to have been
a source of confusion in the past.

Also some doxygenification of comments and formatting...


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/tech@1380 70805675-57b6-444b-bb00-18ecffcb4be8