Removed use of lookup lambda table from assimilation_wstress.
This simplifies the code on the nitrogen branch quite a bit, there's now only one way of doing assimilation_wstress and it always returns assimilation for the individual (including fpar).

Motivation for getting rid of the lookup lambda table also for when running without nlim is that the optimization isn't really worth it when most people will be running with nlim in the future.

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/nitrogen@1505 70805675-57b6-444b-bb00-18ecffcb4be8