Bugfix in daily bvoc call on nitrogen branch.
phot had not been calculated when phot.adtmm was used, if there was water stress (because on the nitrogen branch we store the PhotosynthesisResult in the Individual instead of in the local phot variable).

Also removed that phot variable since it had quite a large scope and was used both for water stressed and non-water stressed photosynthesis.

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/nitrogen@1585 70805675-57b6-444b-bb00-18ecffcb4be8