First sketch of new system for dealing with fluxes.
This also fixes the NPP consistency problem which this branch was created for (by treating npp and annual bvoc fluxes like the other fluxes, stored in the Fluxes class instead of in Individuals (which can die before outannual))

The idea here is to let the Fluxes class deal with tedious book keeping stuff these fluxes have in common. The 'scientific' parts of the code simply report fluxes as they are generated, without caring how it is stored.

Fluxes reported by an Individual, via a member function in Individual, will always check the alive flag so it isn't forgotten.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/npp_consistency@1861 70805675-57b6-444b-bb00-18ecffcb4be8