Replaced the global seed variable with Stand/Gridcell specific seeds.
Done in the same way as in the crop branch, except that there's now also
a seed per Gridcell.

This makes it a bit easier to compare runs in some situations, for instance
two different runs with different land cover setups, or to verify that the
restart functionality is working (since the seed is now being serialized).


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/tech@1969 70805675-57b6-444b-bb00-18ecffcb4be8