Fixed serialization problem with Historic.
The old serialize function tried to serialize without access to private members,
so with only size(), operator[] and add(). This recreates a Historic object
with the same history (and the same order as far as functions like operator[]
and to_array are concerned), but represented in a different order. So for
instance the sum() function, which simply goes through the buffer from start
to end won't give exactly the same results (due to limited floating point
precision).


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/nitrogen@2112 70805675-57b6-444b-bb00-18ecffcb4be8