Reverted r2075 for now.
r2075 seems to change the results, probably due to the fact that the gpterm function used in the new formulation of f() doesn't return negative values. These changes are perhaps safe, global numbers look similar but there are some changes in PFT distribution.

I'm reverting this for now to avoid delaying reintegration of this branch.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/canexch_simplification@2188 70805675-57b6-444b-bb00-18ecffcb4be8