Made the submit script executable.
Before, it used to be started with 'sh submit.sh', which assumed sh was linked
to bash (as it traditionally has been on Linux, but cannot safely be assumed).

Now it can be started with ./submit.sh

sh submit.sh will still work on those systems where it used to work.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/bugfix@2784 70805675-57b6-444b-bb00-18ecffcb4be8