Removed double function of crop grass pft:s by adding dedicated intercrop grass pft:s.
New pft:s CC3G_ic and CC4G_ic are used for intercrop grass growth, while CC3G and CC4G can still be used for pasture grass within the CROPLAND landcover as previously. However, it is recommended to grow pasture grass in the PASTURE landcover, with proper competition between c3 and c4 grasses, while reserving CC3G and CC4G for situations where c3 and c4 monocultures are desired.

Output of pasture grass and intercrop grass is now separated and correct. The intercrop values are the average for all crop stands and are more meaningful if pasture grass is grown in the PASTURE landcover.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/crop@2815 70805675-57b6-444b-bb00-18ecffcb4be8