Moved land cover input code to InData.h and InData.cpp. (Not yet for demo input).
The usage of this code, which is provisional and not subject to the formal format restrictions of main code, can be turned off by defining DYNAMIC_LANDCOVER_INPUT in guess.h. In that case, static equal land cover and crop fractions are enforced. Also, sowing and harvest dates input is disabled, enforcing dynamic calculation of these dates.

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/crop@2839 70805675-57b6-444b-bb00-18ecffcb4be8