Changed the value of tcmax_est for TeBE (back) to 18.8 (from 10).
The value was given as 18.8 in Sitch et al. (2003). It had been changed to 10 at some point before version control, no one knows for sure why but we (me and Paul Miller) believe it was probably a mistake. 

It was spotted by Matthew Forrest, who noted that there was a dead zone in the subtropics where no trees could grow. The reason was that tcmax_est for TeBE was 10, and for tropical trees tcmin_est was set to 15.5. So no trees could establish if the mean temperature of the coldest month was between 10 and 15.5 degrees. This was clearly visible in very southern China and the very south of the States, and probably in the southern hemisphere too.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/parameterisation@2840 70805675-57b6-444b-bb00-18ecffcb4be8