Code updates mostly concerning compatibility between crop and natural vegetation lai terminology.
Crop lai variables consistent with natural veg. lai. Daily crop values now in lai_daily, while lai (only used for printout) is the year's maximum value.

fpc is now set in allometry(), as for natural vegetation. Default is 1.0, but the usual derivation from maximum lai is also working (needs to be defined in the code), seemingly giving slightly higher production, but this option needs to be verified before using.

Some code preparing the nitrogen code for crops, dealing with the daily nature of e.g. cmass_leaf (this code is commented out).



git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/crop@2859 70805675-57b6-444b-bb00-18ecffcb4be8