Two bugfixes in Platon's submit script to avoid error messages.
The check for an old 'local' symlink was broken and gave error messages.

mpiexec now starts as many processes as requested, not as many as is allocated
to the job. This should typically be the same if the recommendation to
request a multiple of 8 is followed, but otherwise you would get error messages
before.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/bugfix@2932 70805675-57b6-444b-bb00-18ecffcb4be8