Reverting changes in previous commit. Burning of unharvested grains is better done by scaling with a burned fraction.
This might be implemented in the future if preferred. Then either burning or residue removal can be prescribed for a crop site, resulting in some differences in carbon fluxes and pools. At the moment res_outtake means residue removal in crop stands and burning in natural stands, probably representing common practices.

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/crop@2993 70805675-57b6-444b-bb00-18ecffcb4be8