Reintegrated tech branch (r2999 - r3000)
Only renamed two files, guessio_cru.cpp was renamed to cruinput.cpp and
guessio.cpp was renamed to demoinput.cpp.

Synchronize with this revision separately to avoid problems.

Since there are no changes _in_ the files, if you get a conflict in for 
instance guessio_cru.cpp, simply replace the new cruinput.cpp with your
old guessio_cru.cpp and resolve the conflict by letting SVN remove the
old file. 

(It's of course not as easy as that if you bring in several revisions
from trunk at once)

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/trunk@3001 70805675-57b6-444b-bb00-18ecffcb4be8