Condensing the parameter list in the harvest functions.
A struct, Harvest_CN, containing the input and output variables used in the harvest functions, is now replacing the individual, patchpft and patch variables previously used in the parameter list. These functions are used directly at land cover change.

Accompanying functions take an individual as a parameter, copy the relevant variables to a Harvest_CN struct and call the above functions. The updated variables are then copied to the individual and patchpft and the patch-level harvest flux is updated. These functions are used in growth() for normal harvest.

This update includes a bug in harvest_natural()/harvest_wood() that resulted in nitrogen storage variables not being updated.

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/crop@3093 70805675-57b6-444b-bb00-18ecffcb4be8