get_yearly_data() now fills the end of an incomplete year differently.
It used to fill the remainder of an incomplete year with zeros. These days
wont be simulated, but the values (at least if monthly) can be used in
interpolations, so it should be better to use something reasonable
(0 K temperature isn't very good). So now the last value is used throughout
the year instead.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/cfinput@3257 70805675-57b6-444b-bb00-18ecffcb4be8