Merged with forest_management branch, getting r3294. Possibility to let pft:s assigned to a specific landcover grow in other landcovers.
Implemented for FOREST, where natural pft:s can grow if NATURALPFTSINFOREST is defined in config.h (undefined by default).
The rules are implemented in create_stand_lu(), where the preferred pft:s in each stand are set to active (in standpft).

Other implemented changes necessary for this option are:

photosynthesis() now takes alphaa, defined by the function alphaa() as a function parameter. This also applies to assimilation_wstress().

The slow C pools and fluxes from these are now printed in the files for the receiving landcovers, not the donating landcovers as before.
The earlier code is left intact but commented out.

Landcover-specific output files should be used for pft-specific output (e.g. anpp). This is not yet implemented, but could be done as for cflux and cpool output.

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/crop@3296 70805675-57b6-444b-bb00-18ecffcb4be8