Bugfix in how we calculate mean forest floor PAR levels during growing season.
We used to only consider leaf phenology, which meant some sites with evergreens
could have a growing season during the whole year. That then meant that PAR 
levels during polar night for instance were included in the mean value, which
should limit establishment (due to the parff_min parameter).

In practice it doesn't seem like the model is very sensitive to parff_min, so
the results don't seem to change much, but this should still be a more correct
way of calculating mean forest floor PAR during growing season.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/bugfix@3353 70805675-57b6-444b-bb00-18ecffcb4be8