Update to crop fraction data input code.
Input file may now omit unused crop pft:s.
If a name of a crop pft with include == true is not found in the input file header, a warning is displayed and that pft is removed from the pft list.
Crop cftid now does not have to be the same as the corresponding input file data column number.

Remaining constraints are:
Crop cftid must be a unique integer between (including) 0 and NCROPSTANDS_MAX-1.
Maximum number of data columns in input file is NCROPSTANDS_MAX.

NB. this update requires an input file with header.
This update does not apply to the landcover input (file with landcover types).

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/crop@3380 70805675-57b6-444b-bb00-18ecffcb4be8