Merged with forest_management branch, getting r3630 and 3631.
Functionality to force winter or spring sowing for wheat and rapeseed. Introducing fallow into crop rotation.

The pft parameter forceautumnsowing (0=off) bypasses calculation of choice between winter/spring sowing and forces winter (1) or spring (2) sowing.

A new st parameter fallow1, fallow2 etc. sets a crop rotation item to fallow. A crop pft name still has to be named in crop1/2 etc. but it will not grow.
Example of a crop rotation in the ins-file:
crop1 "TeWW"
crop2 "TeCo"
crop3 "TeCo"
fallow3 1

The method may have problems caused by the set date for going to a new rotation item after fallow (day 180/364 in N/S hemisphere), e.g. causing out-of-synch jumps in the rotation. Also, while the selected crop pft in fallow should not affect the results, it does in some locations. The recommendation is to check the results for signs of problems for any site and improve the code if needed.

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/crop@3632 70805675-57b6-444b-bb00-18ecffcb4be8