Reintegrated tech branch (r3574 - r3645)
These are mostly changes which make it easier to maintain the branches which couple LPJ-GUESS to some other model, such as a global or regional climate model.

LPJ-GUESS is now better prepared for running with leap years. LPJ-GUESS still runs with a fixed 365 day year, but the code no longer needs to rely on that fact. Instead of using the constant 365 in various places, you can now use Date::MAX_YEAR_LENGTH or date.year_length().

The code in framework() for simulating a single day for a single grid cell has been lifted out into a separate function (makes it easy to reuse this code when the rest of the framework is very different, such as when coupling).

benchmarks: Identical results


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/trunk@3649 70805675-57b6-444b-bb00-18ecffcb4be8