Update of input module structure. Cycling of the last years' climate after the historical data period. Cropland fraction ramp from 0 to first historical value.
Landcover and climate input modules and CO2 input are separated and members of the Input class. Soil code and nitrogen deposition input can be used outside of the climate module, but are kept inside if SOIL_INPUT_IN_CLIMATE_MODULE and NDEP_INPUT_IN_CLIMATE_MODULE are defined, to keep compatibility with benchmarks (temporarily).

Coord is defined in inputdefinitions.h and is also accessible to text input code (currently InData). 
The gridlist and date loops are located in the Input object.

The first and last simulation years are determined by:
1. Setting firsthistyear and lasthistyear in the instruction file.
2. Setting firsthistyear and nyear_hist in the instruction file.
3. Setting nyear_hist in the ins-file (the climate input first historic year obtained by climate_input_module->getfirsthistyear() will be used)
If none of these parameters are set in the instruction file, the first historic year and nyear_hist (obtained by climate_input_module->getnyear_hist())are used.

In cruinput, if firsthistyear occurs before the first historic year in the climate input, the spinup is extended until that year. If lasthistyear occurs after the last historic year in the climate input, the last NYEAR_FUTURE_DATA years of the historic climate are repeatedly cycled (temperature de-trended).

If firsthistyear occurs before the first historic year in the landcover input, the first year's value is used. If lasthistyear occurs after the last historic year in the landcover input, the last year's value is used.
A linear increase from 0 to the first year's cropland fraction value is done for a period of nyears_cropland_ramp years (set in the instruction file). This is avoided by a value of 0 (default).

The cfinput module is not yet updated to work with this verion !

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/crop@3896 70805675-57b6-444b-bb00-18ecffcb4be8