cfinput now functional with new input structure. Plus a few other updates in text file input code.
When using the cfinput module, either a gridlist file with cf reference coordinates can be used (file_gridlist_rc != "") or a standard gridlist file with cru coordinates (file_gridlist).
NB. This code assumes a coordinate resolution of 0.5 degrees (offset of 0.25 degrees between cru and cf coordinates) and needs to be updated to be general !

A few changes to missing data warnings for landcover input files.

Text input can now map data on disk for quicker access if there is a problem with loading all data into memory.

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/crop@3939 70805675-57b6-444b-bb00-18ecffcb4be8