Extended support for different spatial resolutions in data files.
Fix for creating memory copy of text file (LUTOMEMORY defined) when searching for finer spatial resolution coordinates than in the file.
cfinput now supports searching for finer spatial resolution coordinates than in the file.
Added searchradius to cfinput, needed when searching for finer spatial resolution coordinates than in the file.
soilparameters are now searched for in the cru-file with the correct offset, so searchradius can be 0 normally.

The parameter search_for_centre_of_gridcell is set to false by default to make sure the user checks that the spatial resolution settings are correct (spacial resolution of text grid list and data files are parsed, but the success depends on whether the files contain enough coordinates).

NB. The support for lon-lat grid lists with finer spatial resolution coordinates than in the cf file precludes the saving of a cf-format gridlist in memory created by using get_index_for_coords(). get_index_for_coords() must be used for each single search for a coordinate in the cf file.

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/crop_reintegration@4093 70805675-57b6-444b-bb00-18ecffcb4be8