Reverting input structure to trunk (crop pre-r3896).
Version 1: Compared to the version before this revision, this version does not support lon-lat gridlists for the cfinput module.

For input data with different spatial resolutions (supported by cruinput in this revision), the global parameter search_for_centre_of_gridcell must be set to true in the code (default false).

textured_soil is set to true by default.

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/crop_reintegration@4107 70805675-57b6-444b-bb00-18ecffcb4be8