Added instruction (.ins) file crop_n.ins and modified global.ins and the guess.ins files of the two crop benchmarks crop_global and crop_mixed_sites.
The ins files were modified in order to run the crop_global benchmark

Only small changes to global.ins and to the two guess.ins files.
The two guess.ins files points to data files in a data catalog that is temporary - it will change path in the future on Simba.

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/crop_reintegration@4139 70805675-57b6-444b-bb00-18ecffcb4be8