Aet_ifalive_fix applied manually to canexch.cpp
This fix handles the issue that if any individuals in a patch were not alive the interception would be reset to zero for that patch. 

Manually applied checking against both Toms patch and Bugfix branch r4078 (and Michael's luma branch code).

(From Tom, comment about Developers meeting Aug 2014:
As discussed at Tuesday’s developers‘ meeting, there is a small bug in the calculation of evapotranspiration and interception in the trunk version of LPJ-GUESS (and therefore presumably other branches also), whereby individuals which are not alive (indiv.alive==0) are included. Water intercepted and transpired by these non-living individuals affects the soil water content and hence other individuals in the patch. It appears to be particularly important in arid tropical regions, where C3 grass transpires despite not being able to survive. This doesn’t appear to have a notable effect on the global dominant PFT distribution, but soil water content and NPP in arid areas is changed substantially (up to several 10s of %).)

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/crop_reintegration@4208 70805675-57b6-444b-bb00-18ecffcb4be8