First commit of two or more to fix a bug preventing stating a 2nd simulation inside guesswin (the graphical interface)
Fixed here: cru-based simulations
Remains: simulations using the demo- or cf- input modules

Fix includes that gridlist.killall() is invoked before 1st gridcell is read rather than at the end of the simulation in the input-module destructor.
Similarly, resets of pftlist etc are done at the *start* of the simulation.

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/bugfix_crop_reintegr_c@4486 70805675-57b6-444b-bb00-18ecffcb4be8