Fixed missing 'relative' paths to landuse data (for the benchmark version symlink system).
The fixed problem: 
In the .ins file for the two benchmarks which depend on land-use data (crop_global and crop_mixed_sites), the path to the land-use data was hard-coded to the specific simba directory, not defined as a symlink-modifiable path as in the paths.ins.

New paths tested on Simba and works.

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/tech_benchmarks@5416 70805675-57b6-444b-bb00-18ecffcb4be8