Refactored (rename) speciated monoterpene variables (amt1/2, mmt1/2) to more self explanatory names in commonoutput.
Renamed variables of {a,m}mt1 and {a,m}mt2 category to {a,m}mon_endo and {a,m}mon_nonendo, respectively, for the endcoyclic and the non-endocyclic species(the rest), respectively.

Also removed the amt_columns descriptor, assuming changes to the columns of the speciated-MT files will closely follow changes to the amon-file layout, and thus a specific amt_columns descriptor is not needed.

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/tech_bvoc@6209 70805675-57b6-444b-bb00-18ecffcb4be8