Correction of previous revision: renamed speciated monoterpene variables as to be harmonaized with scientific_description.docx.
Renamed variables of r6027 {a,m}mt1 and {a,m}mt2 category to {a,m}mon_mt1 and {a,m}mon_mt2, respectively, for the endcoyclic and the non-endocyclic species(the rest), respectively. This clarifies that mt1 and mt2 a re subgroups of mon (but removes the info about endocyclic nature of subgroup mt1. Instead see scientific_description.docx for that.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/tech_bvoc@6216 70805675-57b6-444b-bb00-18ecffcb4be8