Reverted 4 commits 6454,6369,6368,6332 related to benchmark fluxnet.cpp
Further fluxnet-related files and -folders in catalog benchmarks/ will be reverted/deleted in other commits thatn this one.

Reason: benchmark-fluxnet-related functionality will be moved to a separate branch of its own.

In this commit, the revert of 6454 gave conflict in commonoutput.cpp. Resolved: theirs chosen.
and the revert of 6332 gave conflict in fluxnet.cpp. Resolved: theirs chosen.

Build and small run of 1 gc works.

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/daily_grass@6657 70805675-57b6-444b-bb00-18ecffcb4be8