Added reading of monthly relative humidity from a fastarchive that also contains scalar wind: cruncep_1901_2015wind.bin.
The major news is the addition of the new array double hist_mrhum[NYEAR_HIST][12] in cruinput.h,
and the update of the header file cruncep_1901_2015wind.h (a temporary fastarchive*) to include also relative humidity (rhum).

The model code will find the rhum data in hist_mrhum[NYEAR_HIST][12], by calling CRUInput::getgridcell() as ususal.

*) Decision by Ben is that wind will be included in the misc fastarchive at the timepoint when fastarchives will be generated from cruncep v9.
The same will therefore be applied for this new variable relative humidity.

For now, the cruncep_1901_2015wind.bin is assumed to reside on the same path as cruncep_1901_2015misc.bin,
cruncep_1901_2015wind.bin is for now required for the model to run when firemodel == BLAZE.

The code Compiles and builds (if disregarding compile/build errors from files blaze.cpp, simfire.cpp, and somdynam.cpp.)

It is not tested whether the code runs without error. If errors when reading the wind data, most likely culprit will be the code for derivation of file path for file_cru_wind.

git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/fire_blaze_add_windfastarchive@6729 70805675-57b6-444b-bb00-18ecffcb4be8