Slimmed fix of grass-age bug, in get_stand_age_structure(), commonoutput.cpp; aka bug preventing grass pft before tree pfts in ins file.
This is a reintroduction of the main change of the revision r6446 which was reverted in r6733 of this branch.

The fix has been verified to prevent runtime crash on Windows with 20 gridcell locations, and europe.ins/europe.ins (see further down).
And also, on aurora, crop_global benchmark with a full-global gridlist (reason see last lines in this message).

Explanation to fix:

The bug gave run-time crashes upon some specific additions of pfts to the ins files, or upon setting age>0 for grasses.
Its reason was the following:
The size of allocated memory was:
  densindiv = new double[npft*nageclass];			outannual() in commonoutput.cpp r1471
But writing to it was outside allocated size (since c>0):
  densindiv [pft.id*nageclass + c]	(c < nageclass)		get_stand_age_structure() in commonoutput.cpp r483
(and also reading from it is outside allocated size
  densindiv [pft.id*nageclass+c]	(c < nageclass)		outannual() in commonoutput.cpp r1488)

The present fix solves the problem (the bug above)

Runtime verification tests:

* Added pft copy of BES to end of europe.ins instruction file.
  With fix above: no runtime crash.
  Without fix above: runtime crash at delete[] densindiv_matrix; in outannual() in commonoutput.cpp r1492
* Added code indiv.age++; on row 1458 in growth() in growth.cpp
  With fix above: no runtime crash.
  Without fix above: runtime crash at delete[] densindiv_matrix; in outannual() in commonoutput.cpp r1492
* When instead adding 1 grass at end of europe.ins instruction file, and then 1 or 2 tree pfts: 
  With fix above: no runtime crash.
  However, without fix, the runtime crash appears later than above, *after* 1 gc has been completed). 
  This means that the error can appear randomly, therefore a full-global run was done on Aurora (next bullet)
* Added copy of pft C3_gr and copy of pft BNE to end of global.ins instruction file, and indiv.age++; in growth() as above,
  and ran benchmark crop_global on aurora, with fix above: no runtime crash. 

These tests were done on tech_fixes branch r6446, and the last-bullet test on branch tech_fix_grassage_bug r6735. The present commit is a slimmed version of those two revisions, which were identical, and which in addition to the bugfix here also contained a refactoring of an array variable name.


git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/tech_fixes@6763 70805675-57b6-444b-bb00-18ecffcb4be8